refactor(transfer): extract shared fetch_full_column_types
